Add files via upload
This zip file is the input data from an old drive, so it's as close to the original data as I can get.  The files ending _bin are the binaritized versions.  Note that later researchers already discovered that I was inconsistent in my binaritization at the median.  (As far as I can recall, I binaritized the data as I collected it, and initially I did it by hand in a spreadsheet and only later wrote a program to do it automatically-- and consistently.)